pyproject.toml
[tool.poetry]
name = "pythoneda-artifact-event-versioning"
version = "0.0.1a1"
description = "Events for pythoneda-artifact/versioning"
authors = ["rydnr <github@acm-sl.org>"]
readme = "README.md"
homepage = "https://github.com/pythoneda-artifact-event/versioning"
packages = [{include = "pythonedaartifacteventversioning"}]

[tool.poetry.dependencies]
python = "^3.8"
pythoneda-base = "^0.0.1a16"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
